backend/requirements.txt
fastapi>=0.115.8
langgraph>=0.0.10
langchain>=0.3.18
uvicorn>=0.27.1
python-dotenv>=1.0.0
pydantic>=2.0.0
motor>=3.3.2
pymongo>=4.6.1 